% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topodata.R
\name{calculate_slope_aspect}
\alias{calculate_slope_aspect}
\title{This function computes both slope and aspect from a Digital Elevation Model (DEM).}
\usage{
calculate_slope_aspect(dem)
}
\arguments{
\item{dem}{A raster object representing the Digital Elevation Model.}
}
\value{
A list containing two raster objects: "slope" representing the slope and "aspect" representing the aspect.
}
\description{
This function computes both slope and aspect from a Digital Elevation Model (DEM).
}
\examples{
\dontrun{
# Example usage
elevation_raster <- raster("/Users/chinyereottah/Desktop/Mac/SpatialR/data-raw/CanDEM.tif")
result <- calculate_slope_aspect(elevation_raster)
slope_raster <- result$slope
aspect_raster <- result$aspect
plot(slope_raster, main="Slope")
plot(aspect_raster, main="Aspect")
}

\dontrun{
# Example usage
elevation_raster <- raster("/Users/chinyereottah/Desktop/Mac/SpatialR/data-raw/CanDEM.tif")
result <- calculate_slope_aspect(elevation_raster)
slope_raster <- result$slope
aspect_raster <- result$aspect
plot(slope_raster, main="Slope")
plot(aspect_raster, main="Aspect")
}
}
\seealso{
\code{\link[raster]{focal}}, \code{\link[raster]{atan}}, \code{\link[base]{atan2}}
}
\author{
Chinyere Ottah
}
\keyword{spatial}
