% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopographyR-package.R
\docType{package}
\name{TopographyR-package}
\alias{TopographyR}
\alias{TopographyR-package}
\title{TopographyR: It looks at topographical anlysis in R (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChinyereRuth/TopographyR}
  \item Report bugs at \url{https://github.com/ChinyereRuth/TopographyR/issues}
}

}
\author{
\strong{Maintainer}: Chinyere Ottah \email{Chinyere.Ottah@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
